azure/requirements.txt
psycopg2-binary
python-dotenv
pydantic
sqlalchemy
